% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Copeland.2022.hypoxia.flux-package.R
\docType{package}
\name{Copeland.2022.hypoxia.flux-package}
\alias{Copeland.2022.hypoxia.flux}
\alias{Copeland.2022.hypoxia.flux-package}
\title{Copeland.2022.hypoxia.flux: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: William Oldham \email{wm.oldham@gmail.com} (\href{https://orcid.org/0000-0003-3029-4866}{ORCID})

}
\keyword{internal}
