% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flux_measurements}
\alias{flux_measurements}
\title{Measurements for extracellular flux calculations}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 25213 rows and 17 columns.
}
\usage{
flux_measurements
}
\description{
A data set containing the interpolated metabolite concentrations, cell counts,
and evaporation volumes for extracellular flux determinations.
}
\details{
\describe{
\item{cell_type}{
\code{lf} = lung fibroblasts \cr
\code{pasmc} = pulmonary artery smooth muscle cells}
\item{experiment}{
\code{02} = 0.2\% oxygen for hypoxia \cr
\code{05} = 0.5\% oxygen for hypoxia \cr
\code{bay} = molidustat treatment \cr
\code{05-bay} = 0.5\% oxygen plus molidustat}
\item{batch}{a group of biological replicates analyzed similarly}
\item{date}{start date of an experiment}
\item{metabolite}{name of measurement}
\item{abbreviation}{abbreviated metabolite name}
\item{detector}{
\code{fld} = HPLC fluorescence detector \cr
\code{mwd} = HPLC multi-wavelength detector \cr
\code{enzyme} = enzymatic assay \cr
\code{hplc} = OPD-derivatized HPLC detection \cr
\code{lcms} = liquid chromatography-mass spectrometry assay \cr
\code{picogreen} = fluorescence dye labeling of DNA}
\item{type}{
\code{cells} = conditioned medium \cr
\code{empty} = unconditioned medium}
\item{oxygen}{ambient oxygen level}
\item{treatment}{
\code{None} = no treatment \cr
\code{DMSO} = 0.1\% DMSO \cr
\code{BAY} = 10 μM molidustat}
\item{time}{hours}
\item{well}{denotes technical replicates in each experiment}
\item{conc}{measured in number for cells and μM for metabolites}
\item{volume}{measured in mL, extrapolated from evaporation controls}
\item{nmol}{metabolite mass}
}
}
\keyword{datasets}
