% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{growth_rates}
\alias{growth_rates}
\title{Growth rate parameters}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 124 rows and 10 columns.
}
\usage{
growth_rates
}
\description{
A data set containing the growth rate and X0 values from linear fitting of
cell count data.
}
\details{
\describe{
\item{cell_type}{
\code{lf} = lung fibroblasts \cr
\code{pasmc} = pulmonary artery smooth muscle cells}
\item{experiment}{
\code{02} = 0.2\% oxygen for hypoxia \cr
\code{05} = 0.5\% oxygen for hypoxia \cr
\code{bay} = molidustat treatment \cr
\code{05-bay} = 0.5\% oxygen plus molidustat}
\item{batch}{a group of biological replicates analyzed similarly}
\item{date}{start date of an experiment}
\item{oxygen}{ambient oxygen level}
\item{treatment}{
\code{None} = no treatment \cr
\code{DMSO} = 0.1\% DMSO \cr
\code{BAY} = 10 μM molidustat}
\item{X0}{cell count at time 0}
\item{mu}{growth rate per hour}
}
}
\keyword{datasets}
